Change mapSec type from u8 to u16 for landmarks and match calls
Updated the mapSec field and related function signatures from u8 to u16 in landmark and match call structures and functions to support a larger range of map section IDs. Also fixed theme mapping in map_name_popup.c for dynamic map sections.

+++ include/pokenav.h
@@ -17,7 +17,7 @@ struct PokenavMonListItem
 struct PokenavMatchCallEntry
 {
     bool8 isSpecialTrainer;
-    u8 mapSec;
+    u16 mapSec;
     u16 headerId;
 };
 
@@ -346,7 +346,7 @@ void PokenavList_ReshowListFromCheckPage(void);
 
 // pokenav_match_call_data.c
 bool32 MatchCall_HasCheckPage(u32 idx);
-u8 MatchCall_GetMapSec(u32 idx);
+u16 MatchCall_GetMapSec(u32 idx);
 bool32 MatchCall_HasRematchId(u32 idx);
 bool32 MatchCall_GetEnabled(u32 idx);
 u32 MatchCall_GetRematchTableIdx(u32 idx);
